init.sh is now universal

+++ web/init.sh
@@ -1,10 +1,29 @@
-sudo ln -sf ~/web/etc/nginx.conf /etc/nginx/sites-enabled/test.conf
-sudo rm -rf /etc/nginx/sites-enabled/default
-sudo /etc/init.d/nginx restart
+#!/bin/bash
+curDir=$(pwd)
+homeEnv='/home/user/StepicWebTech/web'
+testEnv='/home/web'
 
-sudo ln -sf ~/web/etc/gunicorn.conf   /etc/gunicorn.d/test
-sudo /etc/init.d/gunicorn restart
+if [ $curDir == $homeEnv ]; then
+        echo 'home env'
 
-cd ~/web/ask/ask
-gunicorn -b 127.0.0.1:8080 wsgi
+	sudo ln -s /home/user/StepicWebTech/web/etc/nginx.conf  /etc/nginx/sites-enabled/test.conf
+	sudo /etc/init.d/nginx restart
+
+	sudo gunicorn -b 127.0.0.1:8000 hello:app
+elif [ $curDir == $testEnv ]; then
+        echo 'test env'
+
+	sudo ln -sf ~/web/etc/nginx.conf /etc/nginx/sites-enabled/test.conf
+	sudo rm -rf /etc/nginx/sites-enabled/default
+	sudo /etc/init.d/nginx restart
+
+	sudo ln -sf ~/web/etc/gunicorn.conf   /etc/gunicorn.d/test
+	sudo /etc/init.d/gunicorn restart
+
+	cd ~/web/ask/ask
+	gunicorn -b 127.0.0.1:8080 wsgi
+
+else
+        echo 'unknown env!'
+fi
